+++ maps/src/HealpixSkyMapInfo.cxx
@@ -357,10 +357,10 @@ HealpixSkyMapInfo::RingAbove(double z) const
 }
 
 void
-HealpixSkyMapInfo::GetInterpPixelsWeights(const Quat &q, std::vector<size_t> & pixels,
+HealpixSkyMapInfo::GetInterpPixelsWeights(const Quat &q, std::vector<uint64_t> & pixels,
     std::vector<double> & weights) const
 {
-	pixels = std::vector<size_t>(4, (size_t) -1);
+	pixels = std::vector<uint64_t>(4, (uint64_t) -1);
 	weights = std::vector<double>(4, 0);
 
 	double z = q.d() / sqrt(dot3(q, q));
@@ -446,11 +446,11 @@ HealpixSkyMapInfo::GetInterpPixelsWeights(const Quat &q, std::vector<size_t> & p
 	}
 }
 
-std::vector<size_t>
+std::vector<uint64_t>
 HealpixSkyMapInfo::QueryDisc(const Quat &q, double radius) const
 {
 	size_t n = 0;
-	auto pixels = std::vector<size_t>(n);
+	auto pixels = std::vector<uint64_t>(n);
 
 	radius /= G3Units::rad;
 	if (radius >= M_PI) {
